Feature/shopping cart service

+++ apps/public-web/eShop.Web.Domain/Dtos/CouponDto.cs
@@ -4,6 +4,6 @@ namespace eShop.Web.Domain;
 public class CouponDto : Entity<Guid>
 {
     public string CouponCode { get; set; }
-    public decimal DiscountAmount { get; set; }
+    public double DiscountAmount { get; set; }
     public int MinAmount { get; set; }
 }
