+++ services/coupon/src/eShop.CouponService.Application.Contracts/Dtos/CouponDto.cs
@@ -4,6 +4,6 @@ namespace eShop.CouponService.Application.Contracts;
 public class CouponDto : EntityDto<Guid>
 {
     public string CouponCode { get; set; }
-    public decimal DiscountAmount { get; set; }
+    public double DiscountAmount { get; set; }
     public int MinAmount { get; set; }
 }
